# api/docs/swagger.yaml
basePath: /api/v1
consumes:
- application/json
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.AggregateData:
    properties:
      count: {}
    type: object
  models.Color:
    properties:
      code:
        type: string
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      hex:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  models.Composition:
    properties:
      algod:
        maximum: 10000
        minimum: 0
        type: integer
      elast:
        maximum: 10000
        minimum: 0
        type: integer
      hilom:
        maximum: 10000
        minimum: 0
        type: integer
      id:
        type: integer
      lino:
        maximum: 10000
        minimum: 0
        type: integer
      nylon:
        maximum: 10000
        minimum: 0
        type: integer
      polye:
        maximum: 10000
        minimum: 0
        type: integer
      rayon:
        maximum: 10000
        minimum: 0
        type: integer
      rayvis:
        maximum: 10000
        minimum: 0
        type: integer
      tencel:
        maximum: 10000
        minimum: 0
        type: integer
      visco:
        maximum: 10000
        minimum: 0
        type: integer
    type: object
...
  models.Fabric:
    properties:
      code:
        type: string
      color:
        $ref: '#/definitions/models.Color'
      color_id:
        type: integer
      composition:
        $ref: '#/definitions/models.Composition'
      composition_id:
        type: integer
      cost:
        type: number
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      supplier:
        $ref: '#/definitions/models.Supplier'
      supplier_id:
        type: integer
    type: object
...
  models.Resource:
    properties:
      code:
        type: string
      color:
        $ref: '#/definitions/models.Color'
      color_id:
        type: integer
      cost:
        type: number
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      supplier:
        $ref: '#/definitions/models.Supplier'
      supplier_id:
        type: integer
    type: object
...
  models.ResourceCreateBody:
    properties:
      code:
        type: string
      color_id:
        type: integer
      cost:
        type: number
      name:
        type: string
      supplier_id:
        type: integer
    required:
    - code
    - color_id
    - cost
    - name
    - supplier_id
    type: object
...
  models.Supplier:
    properties:
      brand:
        type: string
      code:
        type: string
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      nit:
        type: string
      updated_at:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
  title: Vytex API
  version: "0.0"
paths:
...
  /fabrics:
    get:
      description: Get all the fabrics, limit for query o default limit
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Fabric'
            type: array
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Get fabrics from db
      tags:
      - Fabrics
    post:
      description: Create a new fabric
      parameters:
      - description: Fabric create values
        in: body
        name: models.FabricCreateBody
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Fabric'
        "400":
          description: Bad Request
        "409":
          description: Conflict
        "500":
          description: Internal Server Error
      summary: Create fabric
      tags:
      - Fabrics
  /fabrics/aggregate:
    get:
      description: Get aggregate from fabrics
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AggregateData'
            type: array
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Get aggregate from fabrics
      tags:
      - Fabrics
  /fabrics/fabricId:
    get:
      description: Get a fabric by id
      parameters:
      - description: Fabrics ID
        in: path
        name: fabricId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Fabric'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Get a fabric
      tags:
      - Fabrics
...
  /resources:
    get:
      description: Get all the resources, limit for query o default limit
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Resource'
            type: array
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Get resources from db
      tags:
      - Resources
    post:
      description: Create a new resource
      parameters:
      - description: Resource create values
        in: body
        name: resource
        required: true
        schema:
          $ref: '#/definitions/models.ResourceCreateBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Resource'
        "400":
          description: Bad Request
        "409":
          description: Conflict
        "500":
          description: Internal Server Error
      summary: Create a resource
      tags:
      - Resources
  /resources/{id}:
    get:
      description: Get a resource by id
      parameters:
      - description: Resource ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Resource'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Get a resource
      tags:
      - Resources
  /resources/{resourceId}:
    patch:
      description: Update a resource by id
      parameters:
      - description: Resource ID
        in: path
        name: resourceId
        required: true
        type: string
      - description: Resource update values
        in: body
        name: models.ResourceUpdateBody
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Resource'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Update a resource
      tags:
      - Resources
  /resources/aggregate:
    get:
      description: Get result of aggregate function from resources
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AggregateData'
            type: array
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Get aggregate from resources
      tags:
      - Resources
...
  /users:
    get:
      description: Get all the users, limit for query o default limit
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Get users from db
      tags:
      - Users
    post:
      description: Create a new user
      parameters:
      - description: User create values
        in: body
        name: models.UserCreateBody
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Create user
      tags:
      - Users
  /users/{userId}:
    post:
      description: Updates the fields from user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: User update values
        in: body
        name: models.UpdateUserBody
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Update use
      tags:
      - Users
  /users/aggregate:
    get:
      description: Get result of aggregate function from users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AggregateData'
            type: array
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Get aggregate from users
      tags:
      - Users
  /users/me:
    get:
      description: Get the user who do the request with access token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Get the curren user loggged
      tags:
      - Users
  /users/userId:
    get:
      description: Get an user by its ID
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Get a given user
      tags:
      - Users
swagger: "2.0"